requirements.txt
# DocLens - Document Recognition SaaS
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0

# OCR & Document Processing
pytesseract==0.3.10
easyocr==1.7.1
pillow-heif==0.12.0
PyMuPDF==1.24.12

# Image processing
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# Database (async)
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Storage
boto3==1.35.81
python-magic==0.4.27

# Monitoring
prometheus-client==0.21.1

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
